temperature_API/requirements.txt
flask==3.1.1
paho-mqtt==2.1.0
pymongo==4.13.2
